source code cleanup
git-svn-id: https://findbugs.googlecode.com/svn/trunk@12687 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3

+++ findbugsTestCases/src/java/BadIntegerOperations.java
@@ -39,9 +39,9 @@ class BadIntegerOperations {
     @ExpectWarning("BIT")
     int getBytesAsIntUsingOr(byte b[]) {
         int stamp = b[0];
-        stamp |= (int) b[1] << 8;
-        stamp |= (int) b[2] << 16;
-        stamp |= (int) b[3] << 24;
+        stamp |= b[1] << 8;
+        stamp |= b[2] << 16;
+        stamp |= b[3] << 24;
         return stamp;
 
     }
@@ -49,9 +49,9 @@ class BadIntegerOperations {
     @ExpectWarning("BIT")
     int getBytesAsIntUsingPlus(byte b[]) {
         int stamp = b[0];
-        stamp += (int) b[1] << 8;
-        stamp += (int) b[2] << 16;
-        stamp += (int) b[3] << 24;
+        stamp += b[1] << 8;
+        stamp += b[2] << 16;
+        stamp += b[3] << 24;
         return stamp;
 
     }
